Add UniProtParser
Confirmed that it will produce exactly the same output as the version
from ensembl.git, modulo timestamps, once fixes currently residing in
feature/refactor_BaseAdaptor have been merged in. At present there is one,
expected discrepancy - descriptions of dependent xrefs are set to an empty
string instead of inheriting that of the primary (which is what happens
in case of direct xrefs).