UniProtParser::Loader: implement time-based buffer flushing
The scenario: we're parsing TrEMBL xrefs for human with batch size of
1000. There are currently 150,062 TrEMBL records for human and all of
them appear fairly early in the dat file. The parser pushes xrefs
corresponding to 150,000 of these, then sits on the remaining 62 for
several hours until the parser either reaches the end of the file or
aborts for some reason.

Add some code to track how many seconds have elapsed since the last
flush (or, the first time, since the Loader object was constructed) and
call flush() if that time has exceeded specified time - set to default
to 5 minutes.