UniProtParser::Transformer: fix extraction of protein IDs
When I wrote that bit of code I erroneously passed the main id of an
crossreference to the protein-id maker, which was wrong. Now it takes
the whole optional_info arrayref, of which the EMBL/ChEMBL
implementation uses the first element.

Additionally, remove the FIXME notes for the above because with correct
input, both versioned protein IDs and a - in place of one are very much
in use.