UniProtParser::Extractor: streamline the parsing of input lines
To begin with, do prefix filtering before extracting content. There is
no point in applying a capturing regular-expression match to the whole line
only to throw the results away... Moreover, now that we only extract
content after confirming the line at hand is not the record separator we
do not need the content match to be optional.

Secondly, with the prefix (as well as the blanks following it) having
fixed length and seeing as we now process the prefix and the content
separately anyway, do not bother with a regex match here at all. Just
use substr() to extract both the prefixe and the content.

The first change has decreased the exclusive time spent inside
get_uniprot_record() while parsing the Swiss-Prot data file from just
under 1100 to just under 800 seconds, the second one has reduced it
further to around 550 seconds. While this still makes
Extractor::get_uniprot_record() the most time-consuming function in the
whole of a UniProtParser run, at least it is now in the same ballpark as
the runners-up: DBI::st::execute() (around 350 s) and
IO::Handle::getline() (around 250 s).