UniProtParser::Extractor: use predeclared split() field separators
According to the profiler Perl doesn't really optimise this and will
simply recompile the pattern every time split() is called in a loop,
even if the pattern is exactly the same every time. This adds up, even
though the time spent on a single qr{} declaration and compilation of
resulting regex is fairly small. Seeing as we have already declared some
regular-expression patterns at the top of the file, declare field
separators there as well. Bonus: the "semicolon optionally surrounded by
whitespace" pattern appears in several different places and now we can
use a single pattern for all of them.