UniProtParser: stop using Readonly
As it was just before this commit, calls to various parts of Readonly
accounted for almost 15 percent of time (as measured with NYTProf)
required by the parser to process the whole Swiss-Prot data file and
insert human entries from it into the xref database. Moreover, the vast
majority of this time was spent in code pertaining to data accessors and
as such this is NOT fixed overhead whose importance would diminish with
increasing input size. This simply will not do, especially considering
all the constants were internal and therefore if someone starts messing
with them, they will only have themselves to blame.