DBASSParser: put the header under control of Text::CSV
That way all data lines are checked for strict-mode violation. Adapted
is_file_header_valid() accordingly so we can still abort if there is
something wrong with the header itself, with a caveat that all checks
are now effectively case-insensitive (which fortunately doesn't matter
in our case) due to the way Text::CSV::header() works.