MIMParser::extract_ti(): fix end-of-record detection
The original pattern worked for when input records were chomped but that
hasn't been the case for quite a while now as unnecessary. Given that
on the one hand in real OMIM files TI never appears at the end of a
record and on the other all tests involving artificially crafted input
had been conducted manually, the fact the pattern hadn't been updated
went unnoticed until this morning when I first ran unit tests of
extract_ti().

As an added bonus, requiring the records to be unchopped means we should
now be able to detect records which got truncated mid-TI.