UniProtParser::Transformer: explicitly set all direct-xref fields
Before the BaseAdaptor refactoring, upload_xref_object_graphs()
implicitly assigned many properties of the current primary xref to all
associated direct xrefs; this is no longer the case. Therefore,
_make_links_from_crossreferences explicitly assigns the accession,
version, label, description, species_id and info_text of the primary
xref (which is now passed to the method as a whole) to all elements
pushed to @direct_xrefs.