MIMParser::parse_ti(): add missing beginning-of-string anchor
Without \A, the match works fine for correct input but treats an
invalid type symbol as a null one. For example, a TI field like this:
	_123456 fnord mcfnordy
would instead of producing an error be successfully parsed as
	( '', '123456', 'fnord mcfnordy' ) .
Conversely, with the anchor in place entries with invalid type symbol
correctly trigger a match failure.