UniProtParser::Transformer: whitelisted crossreference sources are no longer hardcoded
The list is now passed (as an array in order not to confuse the users
with why we internally use a hash) as an argument to the constructor. If
not specified, no crossreference sources are parsed - meaning neither
direct nor dependent xrefs of any sort are created, only the primary
sequence-mapped ones.