UniProtParser::Extractor: prefixes of interest are no longer hardcoded
The list is now passed (as an array in order not to confuse the users
with why we internally use a hash) as two arguments, one for mandatory
prefixes and another for optional ones, to the constructor. Both lists
default to empty so unless at least one of them is specified, nothing
gets parsed.

Furthermore, since it is no longer guaranteed that any of the fields
previously considered mandatory really are mandatory, all field-parsing
methods now return an appropriately typed null response in the event of
a field being absent.