UniProtParser: strip mid-sequence counts + read sequence type
1. In miRBase dumps each sequence line is followed by a number
indicating how many bases have already been shown.  What this is for
given that all lines except the last one must be 60 bases long and that
there is a space every 10 bases, I've got no idea - but there you do.
Strip those counts before concatenating the sequence lines.

2. In the past Transformer hardcoded the primary-xref sequence type to
'peptide', which will not be correct for e.g. miRBase. Fortunately, in
case of both UniProt and miRBase dumps it is possible to quickly
identify the type of sequence in a record by looking at the length unit
in the record's SQ line. Do so, and provide a clear indicator of an
extraction failure. Note that this is presently of limited usefulness
because Transformer is still hard-wired to insert all direct-xref
mappings into translation_direct_xref; this will be addressed shortly.