UniProtParser::Transformer: direct-xref type is no longer hardcoded
Instead of always assuming it to be 'Translation', try to deduce it from
the sequence type of the current primary xref. In the event of having
failed to do so unambiguously (e.g. this can't be done for DNA
sequences), use a run-wide default which is passed to the constructor.
If the default is not set the type of direct xrefs is set to undef,
which should hopefully trigger an error in the loader.