UniProtParser::Extractor: add a FIXME note in _get_taxon_codes()
Turns out some of the tests in there cannot ever fail and what we get
instead of an exception is the malformed taxonomy-DB entry being quietly
ignored. Since most, if not all, UniProt entries only specify one taxon
code, failing the match results in the whole entry being dropped.

How to fix this will depend on whether we want to retain support for
multiple taxon codes per OX line or not, therefore for the time being
we just add a note.