UniProtParser: use foreach in Extractor and Transformer initialisation
Turns out that even with the arrayrefs pointing to lists of prefixes of
interest declared Readonly before calling Extractor::new(), using the
while ... shift  construct inside said function resulted in emptying
the lists. The same would happen with the list of crossreference sources
passed to the Transformer constructor. This doesn't seem to be a problem
when UniProtParser is run as a whole but came to light when I extended
the unit tests to work with individual ETL modules and attempted to use
the same lists to initialise multiple such modules.